+++ cypress/component/basic/TranslatedMessage.vue
